+++ assets/js/components/Dashboard/IntervalPicker.tsx
